% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_approach.R
\name{CUSUM}
\alias{CUSUM}
\title{Détection d'une rupture avec CUSUM}
\usage{
CUSUM(x, threshold, mu0)
}
\arguments{
\item{x}{vecteur (série temporelle) à valeurs réelles}

\item{threshold}{seuil pour décider d'un changement de tendance}

\item{mu0}{moyenne avant la rupture}
}
\value{
\code{cp} le supposé indice du point de rupture \code{Tn} les statistiques de CUSUM
}
\description{
Elle permet de détecter un point de rupture avec CUSUM sous l'hypothèse que l'on connaît la moyenne avant la rupture.
}
\details{
Cette fonction va calculer des sommes partielles et retourner l'indice de rupture tel que la statistique dépasse un certain seuil en premier.
}
\references{
Romano, Gaetano and Eckley, Idris and Fearnhead, Paul and Rigaill, Guillem (2021) Fast Online Changepoint Detection via Functional Pruning CUSUM statistics, arXiv, 10.48550/ARXIV.2110.08205
}
\seealso{
https://github.com/gtromano/FOCuS
}
