% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_approach.R
\name{CUSUMs_test}
\alias{CUSUMs_test}
\title{Détection de ruptures multiples avec CUSUM ou Page-CUSUM}
\usage{
CUSUMs_test(x, threshold, method = c("CUSUM", "Page-CUSUM"))
}
\arguments{
\item{x}{vecteur (série temporelle) à valeurs réelles}

\item{threshold}{seuil pour décider d'un changement de tendance}

\item{method}{statistique à utiliser parmi "CUSUM" et "Page-CUSUM"}
}
\value{
\code{index_break} les indices associées aux ruptures
}
\description{
Elle permet de détecter plusieurs points de rupture avec, au choix, CUSUM ou Page-CUSUM.
}
