% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FOCuS.R
\name{FOCuS}
\alias{FOCuS}
\title{Fast Online Changepoint Detection via Functional Pruning CUSUM statistics}
\usage{
FOCuS(datasource, thres, ...)
}
\arguments{
\item{datasource}{une fonction qui génère des données (cf. \strong{Details})}

\item{thres}{seuil pour décider d'un changement de tendance}

\item{mu0}{moyenne avant la rupture, si connue. Par défaut est \code{NA} et dans ce cas, la moyenne avant changement est estimée itérativement}

\item{grid}{vecteur de valeurs des amplitudes de changement pour activer l'approximation de la grille de FOCuS. La valeur par défaut est \code{NA}, de sorte que, par défaut, FOCuS fonctionne exactement (cf. \strong{Details})}

\item{K}{La valeur de la fonction de coût bi-poids. La valeur par défaut est \code{Inf}. FOCuS ignorera les points aberrants plus grands que \code{K}}
}
\value{
un objet s3 de la classe FOCuSout où :
\describe{
\item{\code{$t}}{est l'indice de détection,}
\item{\code{$changepoint}}{est le point de rupture,}
\item{\code{$Q1}}{est le coût optimal sous forme de d'une équation du second degré par morceaux à la fin de la séquence au moment de la détection.}
}
}
\description{
FOCuS est un algorithme permettant de détecter les changements de moyenne en temps réel. Ceci est réalisé par une mise à jour récursive d'une équation du second degré par morceaux, dont le maximum est la statistique de test CUSUM pour un changement. FOCuS peut être appliqué à des situations où la moyenne avant changement est connue ou inconnue. En outre, FOCuS peut détecter des changements en présence de points aberrants.
}
\details{
\code{datasource} requiert un objet de la classe \linkS4class{function}. Une fonction valide doit fournir un point des données à chaque appel, et ne doit prendre aucun argument supplémentaire. À chaque itération, FOCuS extrait une observation par un appel de fonction, traite la nouvelle observation et répète cette procédure jusqu'à trouver un point de changement. (cf. \strong{Examples}).
}
\examples{
##################
###   online   ###
##################

set.seed(42)
databuffer <- c(rnorm(3e5, 1), rnorm(1e4, 0))

f <- function() {
  out <- databuffer[i]     # simulating a pull from a buffer
  i <<- i + 1
  out
}

i <- 1; FOCuS(f, 18)
}
\references{
Romano, Gaetano and Eckley, Idris and Fearnhead, Paul and Rigaill, Guillem (2021) Fast Online Changepoint Detection via Functional Pruning CUSUM statistics, arXiv, 10.48550/ARXIV.2110.08205
}
\seealso{
https://github.com/gtromano/FOCuS (Propriétaires de l'algorithme)
}
