% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_approach.R
\name{FOCuS_R}
\alias{FOCuS_R}
\title{Détection d'une rupture avec FOCuS (R)}
\usage{
FOCuS_R(x, threshold)
}
\arguments{
\item{x}{vecteur (série temporelle) à valeurs réelles}

\item{threshold}{seuil pour décider d'un changement de tendance}
}
\value{
\code{cp} le supposé indice du point de rupture et \code{Q} la matrice des itérations des triplets
}
\description{
FOCuS0 permet de détecter un point de rupture avec ou sans la connaissance de la moyenne après la rupture.
}
\details{
Cette fonction va effectuer la première étape de l'algorithme FOCuS0, à savoir calculer Qn en fonction de différents mu1.
Puis (seconde étape) chercher la maximum sur tous les mu de Qn et enfin vérifier si la statistique est plus grand qu'un certain seuil.
}
\references{
Romano, Gaetano and Eckley, Idris and Fearnhead, Paul and Rigaill, Guillem (2021) Fast Online Changepoint Detection via Functional Pruning CUSUM statistics, arXiv, 10.48550/ARXIV.2110.08205
}
\seealso{
https://github.com/gtromano/FOCuS
}
