% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_approach.R
\name{naive_test2}
\alias{naive_test2}
\title{Seconde approche naïve}
\usage{
naive_test2(x, K = 1, method = c("wilcox", "KS"))
}
\arguments{
\item{x}{vecteur (série temporelle) à valeurs réelles}

\item{K}{supposé nombre de ruptures}

\item{method}{statistique à utiliser parmi "wilcox" et "KS"}
}
\value{
\code{index_break} les indices associées aux ruptures
}
\description{
Connaissant le nombre de ruptures, cette fonction segemente astucieusement pour la recherche de points de rupture avec, au choix le test de Wilcoxon ou de Kologorov-Smirnov.
}
\details{
A chaque itération, on cherche l'indice associé à la plus petite p-value obtenue dans les sous-séquences données à l'itération précédente.
Attention, on suppose l'existence d'un point de rupture au minimum.
}
