% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_approach.R
\name{pageSeq}
\alias{pageSeq}
\title{Approche séquentielle de Page}
\usage{
pageSeq(x, mu1)
}
\arguments{
\item{x}{vecteur (série temporelle) à valeurs réelles}

\item{mu1}{moyenne après la rupture}
}
\value{
\code{cp} le supposé indice du point de rupture et \code{Qn} les statistiques de la méthode séquentielle de Page (somme des rapports des vraisemblances logarithmiques)
}
\description{
Elle permet de détecter un point de rupture séquentiellement sous l'hypothèse que l'on connaît la moyenne après la rupture.
}
\details{
Cette fonction va calculer des rapports de log-vraisemblances et retourner l'indice de rupture tel que la statistique finale est non nulle (on suppose d'ailleurs que la moyenne avant la rupture est nulle).
}
