% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_approach.R
\name{Q_iter}
\alias{Q_iter}
\title{Itération de l'algorithme FOCuS0}
\usage{
Q_iter(Q, s, n)
}
\arguments{
\item{Q}{matrice de triplets à valeurs réelles}

\item{s}{somme cumulative}

\item{n}{instant de la nouvelle observation}
}
\value{
\code{Q} matrice après itération
}
\description{
Q_iter permet de mettre à jour Q avec un nouveau triplet (tau, s, l).
}
\details{
Cette fonction va effectuer une mise à jour de Q par rapport à la nouvelle donnée pour l'algorithme de FOCuS0.
}
\references{
Romano, Gaetano and Eckley, Idris and Fearnhead, Paul and Rigaill, Guillem (2021) Fast Online Changepoint Detection via Functional Pruning CUSUM statistics, arXiv, 10.48550/ARXIV.2110.08205
}
\seealso{
https://github.com/gtromano/FOCuS
}
