% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_approach.R
\name{wilcoxtest}
\alias{wilcoxtest}
\title{Test de Wilcoxon}
\usage{
wilcoxtest(x, y)
}
\arguments{
\item{x}{vecteur (série temporelle) à valeurs réelles}

\item{y}{vecteur (série temporelle) à valeurs réelles que l'on souhaite comparer à x}
}
\value{
\code{p.value} la p-value du test de Wilcoxon
}
\description{
Cette fonction retourne la version corrigée de la p-value du test de Wilcoxon (ou somme des rangs) pour deux séquences non appariées.
}
\details{
Sous l'hypothèse nulle, on suppose que les deux séquences proviennent de la même distribution et que \code{Mediane(Y-X) = 0}.
}
\seealso{
https://github.com/obouaziz/robusTest (Inspiration pour l'implémentation)
}
