% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_approach.R
\name{FOCuS_R}
\alias{FOCuS_R}
\title{Détection d'une rupture avec FOCuS0 (R)}
\usage{
FOCuS_R(Q, mu, X)
}
\arguments{
\item{Q}{matrice de triplets à valeurs réelles}

\item{mu}{moyenne après la rupture}

\item{X}{vecteur (série temporelle) à valeurs réelles}
}
\value{
\code{Q_plus} la matrice des itérations des triplets qui détermine le point de rupture \code{S} les statistiques de FOCuS0
}
\description{
FOCuS0 permet de détecter un point de rupture sous l'hypothèse que l'on connaît la moyenne après la rupture.
}
\details{
Cette fonction va effectuer la première étape de l'algorithme FOCuS0, à savoir calculer Qn en fonction de mu.
}
\references{
Romano, Gaetano and Eckley, Idris and Fearnhead, Paul and Rigaill, Guillem (2021) Fast Online Changepoint Detection via Functional Pruning CUSUM statistics, arXiv, 10.48550/ARXIV.2110.08205
}
\seealso{
https://github.com/gtromano/FOCuS
}
